+++ src/lagrangian/cfdemParticle/subModels/locateModel/engineSearch/engineSearch.C
@@ -84,7 +84,7 @@ label engineSearch::findCell
 (
     double** const& mask,
     double**& positions,
-    double**& cellIDs,
+    int**& cellIDs,
     int size
 ) const
 {
